Add nu-builder Added a POC site builder in Zig, `scripts/nu-builder`.
Right now it only works on page files in the site directory.

It'll index all of the markdown files in the page directory, generate a sitemap, then write an html file that includes the raw markdown for that page along with the sitemap. So, the site is fully navigable, but there's no styling or markdown rendering, including directives parsing like `::partial`.

What it needs now to achieve feature parity with the svelte-kit/remark builder, in no particular order:
- process markdown into html
- call into some arbitrary plugin for html content when a directive is encountered in the markdown
- parse yaml frontmatter according to yaml spec
- minify resulting html
- implement partials engine, to be able to inject repeated content throughout the generated content
- implement collections engine, to be able to index, group, and operate on pages
- implement theme engine, to be able to inject theme CSS into generated html

There are some design goals that I have with the nu-builder. I have speed, scaling, and robustness in mind. I'm trying to follow Tiger Style.

One aspect of this has to do with allocating all necessary memory upfront, to avoid slowdowns due to dynamic allocation. To that end, I want to encourage the user to think about realistic sizes for things. For example, one of the constants in the codebase is `const size_of_alice_txt = 1189000`. It's used as the size of a buffer to read in the contents of a single markdown page. Personally, I don't envision the size of a single markdown page in my site to exceed the text copy of Alice's Adventures in Wonderland. The program will crash if this is the case. I could easily make it not crash by using a StackFallbackAllocator rather than a FixedBufferAllocator, but I'd rather keep a reliable performance bar until the demands of the program are questioned again.