Add `run-benchmark` build command and optimize large site generation
Adds a copy of htmx to the repo.

Also, add build commands
- `generate-benchmark`
- `check`

As-is, this tool can generate a 100,000-page site and a sitemap. These files have almost no content in them, and the generated site takes up 15GB of storage. This amounts to about 157kb of boilerplate per file. I'm still embedding a css file (1kb) and (unminified) htmx (155kb), which collectively amount to 155kb, so the HTML boilerplate is closer to 1kb for near-nothing. The size of the generated site isn't too important right now, but I figured I'd share it for context.

The sitemap generation took some refactoring. One, I was embedding the sitemap _of the entire site_ in each page, which got me generating build folders closer to 200GB in size. So I modified it to use an approach similar to an iframe, where the content is loaded in after the initial document.

I was also reading the sitemap into one, contiguous buffer, rather than dynamic memory, to avoid dynamic allocations. However, there is no good size for the buffer when a site can range from one page to hundreds of thousands of pages. So, I open the file ahead of iterating over the pages and write directly to the sitemap file after building the buffer. And that's also then going through a buffered writer. Now that I'm writing this, there are some inconsistencies. I don't really need a buffer for the sitemap entry html since I am now able to write to the file directly. But, if I do continue to use the buffer, am I already buffering my writes enough? Am I adding more buffers here for not much gain? I feel like it would be wise to get rid of one of the buffers now. In an upcoming commit I'll get rid of the sitemap item buffer.

I also renamed the GPA allocator `unlimited_allocator` so it can pop out in my eye more while I'm scanning the code.

There were also some "process fd limit reached" errors while benchmarking because I had failed to close the directories I was opening. Those are now fixed.