Include compiled qjs module
In order to run the site builder, you need to print something to stdin for the js module to parse otherwise it'll stall forever:

```
printf '# Hello, world.\n\nExample text\n' | zig build run-benchmark
```

I'll need to refactor the js module to operate on some input array buffer and write the output to another array buffer, instead of operating on stdin/stdout.
Alternatively, I can investigate how to override the fd for stdin for the js module. But, I think the first approach is more viable to more optimized execution, since it's likely I'll refactor it to a module and just invoke a function from the module, rather than executing the entire module each time.

I also need to figure out how I want to embed the js module in my project. Right now, the process for regenerating the quickjs bytecode looks like this:

```
cd lib/js-md-module
npm i
npx rollup -c
qjsc -c -o ../remark-bin.h bin.js
```

This requires node and qjsc to be available on my PATH. I could compile qjsc from source and invoke it with the build system. Although this is not strictly necessary, since I could `@embedFile("bin.js")` and just get qjs to compile it at runtime.

That leaves the node + rollup steps before it. I need node/npm to fetch the remark set of packages and all of their dependencies, while I need rollup to bundle my entry file `entry.js` and all of its imports into a single file, which is a requirement for compatibility with qjsc. I think it might be easiest to execute this process once and commit the result to source code. The only time this generated file would need to change is if I want to modify the remark plugins.