Change source.Filesystem into source.filesystem.walker
The zig standard library includes `fs.Dir.iterate()`, which I'm using, to find all files in a directory. However, it doesn't find all files recursively, so I keep a buffer of directory handles to pick up directories as they're found for later iteration. In realizing that there is also `fs.Dir.walk()`, a similar function which produces an iterator that _does_ find all files recursively, I thought to swap out the implementation. However, `fs.Dir.walk` requires an allocator, while the `source.Filesystem` iterator does not. So I decided to rename it to `source.filesystem.Walker`. It accomplishes a similar goal as `fs.Dir.walk` but 1) it only yields files, and 2) it requires no dynamic allocations.