Organize DB statements
Part of sqlite usage requires that the DB select query and struct type have the
same fields, in the same order. The library has a safety check for the
number of fields but does not enforce that the names of the columns and
struct fields match. This can lead to subtle bugs, so I've organized the
DB statements such that the query and return type are more closely
colocated. This way, this type of error can be caught at a glance and a
head-scratching moment may be avoided.