Move `Page` and `Template` tables to the `storage` namespace.
Application code -> storage -> Database -> sqlite

Application code (i.e. the goku cli) stores and retrieves site data at
runtime using the `storage` namespace, which exposes `Page` and
`Template` database tables. The storage layer uses `Database`, which
manages a `sqlite` in-memory database  at runtime and exposes type
generation functions for `Table`s and `Iterator`s. The `sqlite` backend
is a third party tool to manage an in-memory database.

This separation isn't yet perfect, as the declarations in the `storage`
namespace don't provide a way to fetch the data - the
`Database.Iterator` is used for that instead, so `storage` is currently
only used half as much as it should be:

Application code ---------> storage --> Database -> sqlite
 (use Database.Iterator)\___________/

In the future I want to add a helpful
`Database.Table(...).Iterator(DestType)` that constructs the `select`
statement automatically: `select field1, field2, ... from table_name;`.
This would help with using `sqlite.Iterator` directly, where the
destination type fields need to be in the same order as they are
specified in the select statement.