Restructure tests to improve emitted docs
`src/test.zig` renamed to `src/root.zig` to indicate to the build system
that this file should be used as the root module for docs generation.

All referenced files are made public in `src/root.zig` in order for docs
to generate links to each type/namespace; these aren't meant to be
imported through `src/root.zig` in actual code.

> If the system doesn't find a file named "foo/root.zig" or "foo/foo.zig", it will use the first file in the tar as the module root.
>
> See: https://github.com/ziglang/zig/wiki/Contributing/3bd4232c3fd11284a00f871e1a7d4c90667dda3f#the-sources-tarball